:recycle:(refactor): The endpoint was updated, renaming 'filter' to 'metrics'. Additionally, the code was refactored, replacing the filter service and tests with metrics to ensure consistency and standardization.
