+++ src/main/java/com/pardal/app/service/Metrics/MetricsService.java
@@ -1,0 +1,8 @@
+package com.pardal.app.service.Metrics;
+
+import com.pardal.app.entity.Dto.FilterDataDto;
+
+public interface MetricsService
+{
+    public FilterDataDto getFilterData(int page, int pageSize);
+}
