+++ src/test/java/com/pardal/app/service/Metrics/MetricsServiceImplTest.java
@@ -1,11 +1,14 @@
-package com.pardal.app.service;
+package com.pardal.app.service.Metrics;
 
-import com.pardal.app.entity.Company;
-import com.pardal.app.entity.Dto.FilterDataDto;
-import com.pardal.app.entity.Product;
-import com.pardal.app.repository.CompanyRepository;
-import com.pardal.app.repository.ProductRepository;
-import com.pardal.app.service.Filter.FilterService;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoInteractions;
+import static org.mockito.Mockito.when;
+
+import java.util.Collections;
+import java.util.List;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -18,14 +21,15 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 
-import java.util.Collections;
-import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.*;
+import com.pardal.app.entity.Company;
+import com.pardal.app.entity.Product;
+import com.pardal.app.entity.Dto.FilterDataDto;
+import com.pardal.app.repository.CompanyRepository;
+import com.pardal.app.repository.ProductRepository;
 
 @ExtendWith(MockitoExtension.class)
-class FilterServiceTest {
+class MetricsServiceImplTest
+{
 
     @Mock
     private CompanyRepository companyRepository;
@@ -34,7 +38,7 @@ class FilterServiceTest {
     private ProductRepository productRepository;
 
     @InjectMocks
-    private FilterService filterService;
+    private MetricsService metricsSerive;
 
     private Page<Company> companyPage;
     private Page<Product> productPage;
@@ -54,7 +58,6 @@ class FilterServiceTest {
         productPage = new PageImpl<>(productList);
     }
 
-
     @Test
     @DisplayName("Should return FilterDataDto when valid page and size are provided")
     void getFilterData_whenValidPageAndSize_shouldReturnDto() {
@@ -65,7 +68,7 @@ class FilterServiceTest {
         when(companyRepository.findAll(pageable)).thenReturn(companyPage);
         when(productRepository.findAll(pageable)).thenReturn(productPage);
 
-        FilterDataDto result = filterService.getFilterData(page, pageSize);
+        FilterDataDto result = metricsSerive.getFilterData(page, pageSize);
 
         assertNotNull(result);
         assertEquals(companyPage, result.getCompanies());
@@ -75,7 +78,6 @@ class FilterServiceTest {
         verify(productRepository, times(1)).findAll(pageable);
     }
 
-
     @Test
     @DisplayName("Should throw IllegalArgumentException when page number is less than 1")
     void getFilterData_whenPageIsLessThanOne_shouldThrowException() {
@@ -84,14 +86,13 @@ class FilterServiceTest {
 
         IllegalArgumentException thrown = assertThrows(
                 IllegalArgumentException.class,
-                () -> filterService.getFilterData(invalidPage, pageSize),
+                () -> metricsSerive.getFilterData(invalidPage, pageSize),
                 "Expected IllegalArgumentException for page < 1, but didn't throw."
         );
 
         assertEquals("O número da página deve ser maior que 0", thrown.getMessage());
         verifyNoInteractions(companyRepository, productRepository);
     }
-
 
     @Test
     @DisplayName("Should throw IllegalArgumentException when page size is less than 1")
@@ -101,11 +102,12 @@ class FilterServiceTest {
 
         IllegalArgumentException thrown = assertThrows(
                 IllegalArgumentException.class,
-                () -> filterService.getFilterData(page, invalidPageSize),
+                () -> metricsSerive.getFilterData(page, invalidPageSize),
                 "Expected IllegalArgumentException for pageSize < 1, but didn't throw."
         );
 
         assertEquals("O tamanho da página deve ser maior que 0", thrown.getMessage());
         verifyNoInteractions(companyRepository, productRepository);
     }
+
 }
